move parameter traits to detail namespace
