+++ Core/include/Acts/Utilities/ParameterDefinitions.hpp
@@ -9,9 +9,7 @@
 #pragma once
 
 #include "Acts/Utilities/Definitions.hpp"
-#include "Acts/Utilities/ParameterTypes.hpp"
 
-#include <cmath>
 #include <type_traits>
 
 // The user can override the (track) parameter ordering and underlying scalar
@@ -122,120 +120,6 @@ static_assert(eFreePos1 == eFreePos0 + 1u, "Position must be continous");
 static_assert(eFreePos2 == eFreePos0 + 2u, "Position must be continous");
 static_assert(eFreeDir1 == eFreeDir0 + 1u, "Direction must be continous");
 static_assert(eFreeDir2 == eFreeDir0 + 2u, "Direction must be continous");
-
-namespace detail {
-
-template <BoundIndices>
-struct BoundParameterTraits;
-template <>
-struct BoundParameterTraits<BoundIndices::eBoundLoc0> {
-  using type = local_parameter;
-};
-template <>
-struct BoundParameterTraits<BoundIndices::eBoundLoc1> {
-  using type = local_parameter;
-};
-template <>
-struct BoundParameterTraits<BoundIndices::eBoundPhi> {
-  static constexpr double pMin() { return -M_PI; }
-  static constexpr double pMax() { return M_PI; }
-  using type = cyclic_parameter<double, pMin, pMax>;
-};
-template <>
-struct BoundParameterTraits<BoundIndices::eBoundTheta> {
-  static constexpr double pMin() { return 0; }
-  static constexpr double pMax() { return M_PI; }
-  using type = bound_parameter<double, pMin, pMax>;
-};
-template <>
-struct BoundParameterTraits<BoundIndices::eBoundQOverP> {
-  using type = unbound_parameter;
-};
-template <>
-struct BoundParameterTraits<BoundIndices::eBoundTime> {
-  using type = unbound_parameter;
-};
-
-template <FreeIndices>
-struct FreeParameterTraits;
-template <>
-struct FreeParameterTraits<FreeIndices::eFreePos0> {
-  using type = unbound_parameter;
-};
-template <>
-struct FreeParameterTraits<FreeIndices::eFreePos1> {
-  using type = unbound_parameter;
-};
-template <>
-struct FreeParameterTraits<FreeIndices::eFreePos2> {
-  using type = unbound_parameter;
-};
-template <>
-struct FreeParameterTraits<FreeIndices::eFreeTime> {
-  using type = unbound_parameter;
-};
-template <>
-struct FreeParameterTraits<FreeIndices::eFreeDir0> {
-  using type = unbound_parameter;
-};
-template <>
-struct FreeParameterTraits<FreeIndices::eFreeDir1> {
-  using type = unbound_parameter;
-};
-template <>
-struct FreeParameterTraits<FreeIndices::eFreeDir2> {
-  using type = unbound_parameter;
-};
-template <>
-struct FreeParameterTraits<FreeIndices::eFreeQOverP> {
-  using type = unbound_parameter;
-};
-
-template <typename indices_t>
-struct ParametersSize;
-template <>
-struct ParametersSize<BoundIndices> {
-  static constexpr unsigned int size =
-      static_cast<unsigned int>(BoundIndices::eBoundSize);
-};
-template <>
-struct ParametersSize<FreeIndices> {
-  static constexpr unsigned int size =
-      static_cast<unsigned int>(FreeIndices::eFreeSize);
-};
-
-}  // namespace detail
-
-/// Single bound track parameter type for value constrains.
-///
-/// The singular name is not a typo since this describes individual components.
-template <BoundIndices kIndex>
-using BoundParameterType = typename detail::BoundParameterTraits<kIndex>::type;
-
-/// Single free track parameter type for value constrains.
-///
-/// The singular name is not a typo since this describes individual components.
-template <FreeIndices kIndex>
-using FreeParameterType = typename detail::FreeParameterTraits<kIndex>::type;
-
-/// Access the (Bound/Free)ParameterType through common struct
-///
-/// @tparam T Parameter indices enum
-/// @tparam I Index of @p T
-template <typename T, T I>
-struct ParameterTypeFor {};
-
-/// Access for @c BoundIndices
-template <BoundIndices I>
-struct ParameterTypeFor<BoundIndices, I> {
-  using type = BoundParameterType<I>;
-};
-
-/// Access for @c FreeIndices
-template <FreeIndices I>
-struct ParameterTypeFor<FreeIndices, I> {
-  using type = FreeParameterType<I>;
-};
 
 // The following matrix and vector types are automatically derived from the
 // indices enums and scalar typedefs.
